+++ modules/private_service_access/versions.tf
@@ -32,11 +32,11 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:private_service_access/v26.2.1"
+    module_name = "blueprints/terraform/terraform-google-sql-db:private_service_access/v27.2.0"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:private_service_access/v26.2.1"
+    module_name = "blueprints/terraform/terraform-google-sql-db:private_service_access/v27.2.0"
   }
 
 }